pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bilibili-api",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "openai>=1.70.0",
    "python-dotenv>=1.1.0",
    "pyyaml!=5.4.1",
]
